Started moving files around to prepare for headless location service
